+++ src/BeShell.hpp
@@ -31,24 +31,28 @@ namespace be {
 
         BeShell() ;
         ~BeShell() ;
-        void setup();
+        void setup(const char * mainScriptPath=nullptr) ;
         void loop() ;
         void run() ;
-        void main() ;
+        void main(const char * mainScriptPath=nullptr) ;
 
         void addLoopFunction(LoopFunction func, void * opaque=nullptr, bool ignoreRepeat=true) ;
         
         template <typename ModuleClass>
-        void use(const char * name=nullptr) {
-            useModule<ModuleClass>(name) ;
+        void use() {
+            static_assert(std::is_base_of<NativeModule, ModuleClass>::value, "ModuleClass must be a subclass of NativeModule") ;
+            engine->mloader.add<ModuleClass>(this, nullptr) ;
         }
+
         template <typename ModuleClass>
         void useModule(const char * name=nullptr) {
+            static_assert(std::is_base_of<NativeModule, ModuleClass>::value, "ModuleClass must be a subclass of NativeModule") ;
             engine->mloader.add<ModuleClass>(this, name) ;
         }
 
         template <typename DriverClass>
         void useDriver() {
+            static_assert(std::is_base_of<NativeClass, DriverClass>::value, "DriverClass must be a subclass of NativeClass") ;
             be::driver::DriverModule::useDriver<DriverClass>(this) ;
         }
     } ;